fix scan loading from settings file
