+++ ptychodus/model/scan/__init__.py
@@ -3,8 +3,7 @@ from .cartesian import CartesianScanRepositoryItem
 from .core import ScanCore, ScanRepositoryKeyAndValue, ScanPresenter
 from .factory import ScanRepositoryItemFactory
 from .lissajous import LissajousScanRepositoryItem
-from .repository import ScanRepository
-from .repositoryItem import ScanRepositoryItem
+from .repository import ScanRepository, ScanRepositoryItem
 from .spiral import SpiralScanRepositoryItem
 from .transformed import TransformedScanRepositoryItem
